Clarified Help, Stopped command from being so chatty
Clrified the help, added "sub-" to make it more obvious that the new
folder would be a sub-folder.  The command had been sending all the
properties of the command out to the user.  Modified coomand to use "|
Out-Null" like other commands in the module do.  Added the $FolderName
to the Write-Verbose text.