Update Out-RsFolderContent.ps1
Loop for folders first, because we need to ensure folders are written to disk before we
attempt to write any files. Otherwise, file writes will randomly fail to write because
its folder has not been created yet.

The Solution was to loop and create all folders then work on the files.

Basically, create all folders, then write all the files.