Adding new functions for CacheRefreshPlans (#308)
* Added Get-RsCacheRefreshPlan function.

* Added Start-RsReportRefresh function.

* Renamed Get-CacheRefreshPlan file.

* Added function names.

* Added New-RsRestCacheRefreshPlan function

* Added test for New-RsRestCacheRefreshPlan

* Added two more Pester tests.

Added two more Pester tests, corrected name of the Start- function, corrected URL used for testing.

* Updated file names.  Changed to Should Not -Be.

* Changed to Should Not Be.

Changed Should Not -Be to Should Not Be.

* Addressed naming & cosmetic issues

Addressed naming & cosmetic issues mentioned on code review.

* Updated Get-RsRestCacheRefreshPlan for multiple results

Updated Get-RsRestCacheRefreshPlan to better present results when a Power BI report multiple CacheRefreshPlans.

* Added second test for New-RsRestCacheRefreshPlan.

Added extra help examples to show how to use the -StartDateTime and -Recurrence parameters.  Also added second test for New-RsRestCacheRefreshPlan.

* Update Start-RsRestCacheRefreshPlan.ps1

Added logic to skip starting any CacheRefreshPlans when the -RsReport parameter is specified and multiple CacheRefreshPlans are present.

* Update Start-RsRestCacheRefreshPlan.ps1

Small clarification / amplification of the help info.

* Added warning to New-RsRestCacheRefreshPlan

Added a warning message to New-RsRestCacheRefreshPlan in case a users tries to use this command with a non-Power BI report.