Set Inherit Parent Policy to false (#368)
Without setting this property to false, this function will not work on objects that have 'InheritParentPolicy' set to true. It sets to false when granting permissions (in Grant RsRestItemAccessPolicy), but does not set to false in this one. In any case, when changing permissions, granting or revoking, the inheritence should be set to false as per the behaviour of the PBIRS app.